refactor: state private methods and rename `coloc_utils` folder
